Move card enums to constants module
