+++ altered/constants/__init__.py
@@ -1,2 +1,3 @@
+from .card import CardExtension, CardRarity
 from .duration import Duration
 from .faction_colors import FACTION_COLORS
