+++ altered/models/card.py
@@ -1,20 +1,8 @@
 from django.contrib import admin
 from django.db import models
 
+from altered.constants.card import CardExtension, CardRarity
 from altered.constants.faction import Faction
-
-
-class CardExtension(models.TextChoices):
-    BEYONG_THE_GATE = "BEYONG_THE_GATE", "Beyong the gate"
-    TRIAL_BY_FROST = "TRIAL_BY_FROST", "Trial by frost"
-    WHISPERS_FROM_THE_MAZE = "WHISPERS_FROM_THE_MAZE", "Whispers from the maze"
-    SKYBOUND_ODYSSEY = "SKYBOUND_ODYSSEY", "Skybound Odyssey"
-
-
-class CardRarity(models.TextChoices):
-    COMMON = "COMMON", "Commun"
-    RARE = "RARE", "Rare"
-    UNIQUE = "UNIQUE", "Unique"
 
 
 class Card(models.Model):
